install/pyproject.toml
[tool.poetry]
name = "install"
version = "0.1.0"
description = ""
authors = ["233 <ayusummer233@vip.qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gitpython = "^3.1.43"
httpx = "^0.27.2"
colorlog = "^6.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
